move proxy types into separate module
